+++ zerotier-one-api/src/models/status.rs
@@ -22,9 +22,9 @@ pub struct Status {
     #[serde(rename = "online", skip_serializing_if = "Option::is_none")]
     pub online: Option<bool>,
     #[serde(rename = "planetWorldId", skip_serializing_if = "Option::is_none")]
-    pub planet_world_id: Option<i32>,
+    pub planet_world_id: Option<i64>,
     #[serde(rename = "planetWorldTimestamp", skip_serializing_if = "Option::is_none")]
-    pub planet_world_timestamp: Option<i32>,
+    pub planet_world_timestamp: Option<i64>,
     #[serde(rename = "publicIdentity", skip_serializing_if = "Option::is_none")]
     pub public_identity: Option<String>,
     #[serde(rename = "tcpFallbackActive", skip_serializing_if = "Option::is_none")]
